+++ ChangeLog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- removed dependencies on `velocypack-module-jdk8` and `velocypack-module-joda`
+- upgraded dependency `com.arangodb:arangodb-java-driver:6.9.0`
+
 ## [3.4.1] - 2021-01-26
 
 - upgraded dependency `com.arangodb:arangodb-java-driver:6.8.2`